api/requirements.txt
fastapi>=0.115
uvicorn>=0.30
pydantic>=2.7
pyyaml>=6.0
